Co-locate bespoke DnD hooks with their components
