--- packages/dnd/src/genericTests/index.ts
@@ -1,1 +1,0 @@
-export * from './GenericDropTarget';
