+++ packages/dnd/src/index.ts
@@ -2,5 +2,4 @@ export * from './InlineCombinatorDnD';
 export * from './QueryBuilderDnD';
 export * from './RuleDnD';
 export * from './RuleGroupDnD';
-export * from './hooks';
 export * from './types';
