+++ packages/react-querybuilder/src/components/QueryBuilder.useQueryBuilderSchema.ts
@@ -1,6 +1,7 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { LogType, standardClassnames } from '../defaults';
-import { useControlledOrUncontrolled, useDeprecatedProps } from '../hooks';
+import { useControlledOrUncontrolled } from '../hooks/useControlledOrUncontrolled';
+import { useDeprecatedProps } from '../hooks/useDeprecatedProps';
 import { getQuerySelectorById, useQueryBuilderSelector } from '../redux';
 import {
   _RQB_INTERNAL_dispatchThunk,
